python/rules_export/requirements.txt
pan-os-python==1.7.3
pandas==1.3.5
python-dotenv==0.19.2
